// src/enums/Coordinates.ts
 enum Coordinates {
  shipInitialPoint_x = 760,
  shipInitialPoint_y = 290,
  EntryToPortsPoint_x = 270,
  EntryToPortsPoint_y = 290,
  portsPoint_x = 60,
  port1Point_y = 120,
  port2Point_y = 234,
  port3Point_y = 366,
  port4Point_y = 498,
  ExitFromPortPoint_x = 270,
  ExitFromPortPoint_y = 330,
  EntryToQueue_x = 700,
  EntryToQueue_y = 240,
  QueueThirdPoint_x = 600,
  QueueSecondPoint_x = 500,
  QueueFirstPoint_x = 400,
  QueuePoint_y = 240,
  ShipDeletionPoint_x = 760,
  ShipDeletionPoint_y = 330,
  ShipDeletionAfterQueuePoint_x = 760,
  ShipDeletionAfterQueuePoint_y = 150,
}

export default Coordinates;